+++ lib/app/core/widgets/quantity_widget.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:money/app/core/helpers/misc_helpers.dart';
+import 'package:money/app/core/helpers/color_helper.dart';
 import 'package:money/app/core/helpers/string_helper.dart';
 
 /// Formatted text using the supplied currency code and optional the currency/country flag
-class QuantifyWidget extends StatelessWidget {
+class QuantityWidget extends StatelessWidget {
   /// Constructor
-  const QuantifyWidget({
+  const QuantityWidget({
     required this.quantity,
     super.key,
     this.align = TextAlign.right,
@@ -23,7 +23,7 @@ class QuantifyWidget extends StatelessWidget {
       textAlign: align,
       style: TextStyle(
         fontFamily: 'RobotoMono',
-        color: isAlmostZero(quantity) ? Colors.grey.withOpacity(0.5) : null,
+        color: getTextColorToUseQuantity(quantity),
       ),
     );
   }